server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Vikram Karthikeyan",email = "viki.kma085@ggmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.135.0,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "supabase (>=2.28.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "clerk-backend-api (>=5.0.2,<6.0.0)",
    "boto3 (>=1.42.59,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
